Multi-unit listing support for Origin.js
This commit adds support and tests for multi-unit listings in the
marketplace resource.

The key changes use the outstanding offers on a listing to determine the
units available for purchase in the marketplace resource. The main new
function is unitsAvailable().

These changes slow down various offer-related actions, so to keep DApp
performance reasonable, it's important to be able to pull listing and
offer data from the discovery server.